US114355 - Byproduct
Added utility method to simulate an HTTP error for a specific request.

To use, simply wrap the method with this call:

```
const sureToFail = ALClient.simulateHttpError( originalRequest, 404, "Not Found", {} );
```

`originalRequest` can be a simple `Promise.resolve(true)` if an instantaneous failure is required.
If `ALClient`'s `collectRequestLog` property is true, the actual details
of the request will be populated into the error; otherwise, only
placeholder values will be provided.